BUG/MINOR: listener: Fix a possible null pointer dereference
It seems to be possible to have no frontend for a listener. A test was missing
before dereferencing it at the end of the function listener_accept().

This patch fixes the issue #264. It must be backported to 2.0 and 1.9.

(cherry picked from commit ad6c2eac28ed96ec8e8c90bd5ce68794ea6bf000)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>