BUG/MINOR: backend: Fix a possible null pointer dereference
In the function connect_server(), when we are not able to reuse a connection and
too many FDs are opened, the variable srv must be defined to kill an idle
connection.

This patch fixes the issue #257. It must be backported to 2.0

(cherry picked from commit 1dbc4676c657072dfbf89a0b7ce9699367f28b33)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>