BUG/MEDIUM: spoe: Use a different engine-id per process
SPOE engine-id is the same for all processes when nbproc is more than 1. So, in
async mode, an agent receiving a NOTIFY frame from a process may send the ACK to
another process. It is abviously wrong. A different engine-id must be generated
for each process.

This patch must be backported to 2.0, 1.9 and 1.8.

(cherry picked from commit d87b1a56d526568b55ee33b77f77c87455026ae1)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>