BUG/MINOR: stream-int: Process connection/CS errors first in si_cs_send()
Errors on the connections or the conn-stream must always be processed in
si_cs_send(), even if the stream-interface is already subscribed on
sending. This patch does not fix any concrete bug per-se. But it is required by
the following one to handle those errors during synchronous sends.

This patch must be backported with the following one to 2.0 and probably to 1.9
too, but with caution because the code is really different.

(cherry picked from commit 328ed220a8c5211aa8b6f37b982f319cf6b3f3d1)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>