MINOR: stats: Add the support of float fields in stats
It is now possible to format stats counters as floats. But the stats applet does
not use it.

This patch is required by the Prometheus exporter to send the time averages in
seconds. If the promex change is backported, this patch must be backported
first.

(cherry picked from commit 88a0db28aeb45e451d36645612e11cd284cc7ef4)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>