BUG/MINOR: mux-h1: Do h2 upgrade only on the first request
When a request is received, if the h2 preface is matched, an implicit upgrade
from h1 to h2 is performed. This must only be done for the first request on a
connection. But a test was missing to unsure it is really the first request.

This patch must be backported to 2.0.

(cherry picked from commit eec96b5381c9508686035842e5886c6cd9d64fa2)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>