BUG/MINOR: stats: Add a missing break in a switch statement
A break is missing in the switch statement in the function
stats_emit_json_data_field(). This bug was introduced in the commit 88a0db28a
("MINOR: stats: Add the support of float fields in stats").

This patch fixes the issue #302 and #303. It must be backported to 2.0.

(cherry picked from commit 52c91bb72ccf20e75f2f79da6ebe345cb2a768f4)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>