REGTEST: mcli/mcli_show_info: launch a 'show info' on the master CLI
This test launches a HAProxy process in master worker with 'nbproc 4'.
It sends a "show info" to the process 3 and verify that the right
process replied.

This regtest depends on the support of the master CLI for VTest.

(cherry picked from commit cd4827746940fb0d55b5e7d027747d98bb2c5f8a)
Signed-off-by: Willy Tarreau <w@1wt.eu>