BUG/MINOR: mux-h1: Adjust header case when chunked encoding is add to a message
When an outgoing h1 message is formatted, if it is considered as chunked but the
corresponding header is missing, we add it. And as all other h1 headers, if
configured so, the case of this header must be adjusted.

No backport needed.

(cherry picked from commit 7a991a9b83526b5000349b185186b0797d9fe080)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>