BUG/MINOR: mux-h1: Capture ignored parsing errors
When the option "accept-invalid-http-request" is enabled, some parsing errors
are ignored. But the position of the error is reported. In legacy HTTP mode,
such errors were captured. So, we now do the same in the H1 multiplexer.

If required, this patch may be backported to 2.0 and 1.9.

(cherry picked from commit 486498c630a0678446808107d02f94c48fc6722a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>