BUG/MINOR: chunk: Fix tests on the chunk size in functions copying data
When raw data are copied or appended in a chunk, the result must not exceed the
chunk size but it can reach it. Unlike functions to copy or append a string,
there is no terminating null byte.

This patch must be backported as far as 1.8. Note in 1.8, the functions
chunk_cpy() and chunk_cat() don't exist.

(cherry picked from commit 48fa033f2809af265c230a7c7cf86413b7f9909b)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>