BUG/MINOR: mux-h1: Mark the output buffer as full when the xfer is interrupted
When an outgoing HTX message is formatted to a raw message, if we fail to copy
data of an HTX block into the output buffer, we mark it as full. Before it was
only done calling the function buf_room_for_htx_data(). But this function is
designed to optimize input processing.

This patch must be backported to 2.0 and 1.9.

(cherry picked from commit a61aa544b4b95d1416fe5684ca2d3a0e110e9743)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>