Revert e8826ded5fea3593d89da2be5c2d81c522070995.
This reverts commit "BUG/MEDIUM: mux_pt: Make sure we don't have a
conn_stream before freeing.".
mux_pt_io_cb() is only used if we have no associated stream, so we will
never have a cs, so there's no need to check that, and we of course have to
destroy the mux in mux_pt_detach() if we have no associated session, or if
there's an error on the connection.

This should be backported to 2.0 and 1.9.

(cherry picked from commit 9dce2c53a8e49d43b501c3025d41705d302b1df1)
Signed-off-by: Willy Tarreau <w@1wt.eu>