BUG/MINOR: mux-h2: also make sure blocked legacy connections may expire
The backport of commit 2dcdc2236 ("MINOR: mux-h2: add a per-connection
list of blocked streams") missed one addition of LIST_ADDQ(blocked_list)
for the legacy version. This makes the stream not be counted as blocked
and will not let the connection expire in this specific case.

This fix is specific to 2.0 and must be backported to 1.9 as well.