BUG/MINOR: mux-h2: do not emit logs on backend connections
The logs were added to the H2 mux so that we can report logs in case
of errors that prevent a stream from being created, but as a side effect
these logs are emitted twice for backend connections: once by the H2 mux
itself and another time by the upper layer stream. It can even happen
more with connection retries.

This patch makes sure we do not emit logs for backend connections.

It should be backported to 2.0 and 1.9.

(cherry picked from commit 9364a5fda33a2f591d5e2640249a54af8955fb8b)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>