BUG/MEDIUM: mux-h2: immediately report connection errors on streams
In case a stream tries to send on a connection error, we must report the
error so that the stream interface keeps the data available and may safely
retry on another connection. Till now this would happen only before the
connection was established, not in case of a failed handshake or an early
GOAWAY for example.

This should be backported to 2.0 and 1.9.

(cherry picked from commit cab2295ae71150d6722505945463b3f1d4627e6e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>