BUG/MINOR: mux-h1: Don't set CS_FL_EOS on a read0 when receiving data to pipe
This is mandatory to process input one more time to add the EOM in the HTX
message and to set CS_FL_EOI on the conn-stream. Otherwise, in the stream, a
SHUTR will be reported on the corresponding channel without the EOI. It may be
erroneously interpreted as an abort.

This patch must be backported to 2.0 and 1.9.

(cherry picked from commit 3f21611bddc40099e0fa4b1b196ee3b691fe7c81)
[wt: adjusted context (no traces)]
Signed-off-by: Willy Tarreau <w@1wt.eu>