BUG/MEDIUM: mworker: don't fill the -sf argument with -1 during the reexec
Upon a reexec_on_failure, if the process tried to exit after the
initialization of the process structure but before it was filled with a
PID, the PID in the mworker_proc structure is set to -1.

In this particular case the -sf argument is filled with -1 and haproxy
will exit with the usage message because of that argument.

Should be backported in 2.0.

(cherry picked from commit 677e2f2c35444f4eda93c363946ba473ddc16a88)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>