BUG/MINOR: stream-int: Fix si_cs_recv() return value
The previous patch on this function (36b536d6c "BUG/MEDIUM: stream-int: Don't
loose events on the CS when an EOS is reported") contains a bug. The return
value is based on the conn-stream's flags. But it may be reset if the CS is
closed. Ironically it was exactly the purpose of this patch...

This patch must be backported to 2.0 and 1.9.

(cherry picked from commit e6d8cb1e911267662c92f257f0f42df1ec78d42b)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>