DOC: Add documentation about the use-service action
The use-service action may be used in tcp-request and http-request rules. It was
added to customize HAproxy reply to a client using an applet (initially a lua
applet). But the documentation was missing.

This patch may be backported as far as 1.6.

(cherry picked from commit 579d83b05f98e061c535537169ce39b7271c3efe)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>