BUG/MINOR: h1: Don't test the host header during response parsing
During the H1 message parsing, the host header is tested to be sure it matches
the request's authority, if defined. When there are multiple host headers, we
also take care they are all the same. Of course, these tests must only be
performed on the requests. A host header in a response has no special meaning.

This patch must be backported to 2.1.

(cherry picked from commit bc7c03eba39ec9f0b94734399853bbece1e1a250)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>