BUG/MINOR: fcgi-app: Make the directive pass-header case insensitive
The header name configured by the directive "pass-header", in the "fcgi-app"
section, must be case insensitive. For now, it must be in lowercase to match an
header. Internally, header names are in lowercase but there is no reason to
impose this syntax in the configuration.

This patch must be backported to 2.1.

(cherry picked from commit bc96c90614b5de59bd8c9d58913111157ed079e7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>