BUG/MINOR: stats: Fix HTML output for the frontends heading
Since the flag STAT_SHOWADMIN was removed, the frontends heading in the HTML
output appears unaligned because the space reserved for the checkbox (not
displayed for frontends) is not inserted.

This patch fixes the issue #390. It must be backported to 2.1.

(cherry picked from commit bc271ec113ab25b2617da73a8b11fce1a56675d4)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>