BUG/MEDIUM: stream: address a very rare unhandled connection issue
Two independent users privately shared reports of haproxy threads or
processes not sleeping in 1.9 showing exactly the same stream state.
These could also be the same as in bug #359 though this is not certain.
Each time the internal state looked like this:

  flags=0x4544e { SF_FINST_D | SF_ERR_SRVCL | SF_REDIRECTABLE | SF_CURR_SESS | SF_BE_ASSIGNED | SF_ADDR_SET | SF_ASSIGNED }
  task=0x2426290 (state=0x204 { TASK_WOKEN_TIMER | TASK_QUEUED } nice=0 calls=396851804 exp=<PAST> tmask=0x1 age=2d22h)
  txn=0x1945af0 flags=0x8003000 { TX_HDR_CONN_PRS | TX_CACHE_COOK | TX_CACHEABLE } meth=1 { HTTP_METH_GET } status=-1 req.st=MSG_ERROR rsp.st=MSG_RPBEFORE
      req.f=0x0c { HTTP_MSGF_VER_11 | HTTP_MSGF_XFER_LEN } blen=0 chnk=0 next=0
      rsp.f=0x00 { 0 } blen=0 chnk=0 next=0
  si[0]=0x20450b8 (state=EST flags=0x280008 { SI_FL_WAIT_DATA | SI_FL_RXBLK_SHUT | SI_FL_RX_WAIT_EP } endp0=CS:0x2176bf0 exp=<NEVER> et=0x000 sub=0)
  si[1]=0x20450f8 (state=EST flags=0x200111 { SI_FL_EXP | SI_FL_ISBACK | SI_FL_NOHALF | SI_FL_RX_WAIT_EP } endp1=CS:0x1a68040 exp=<NEVER> et=0x200 { SI_ET_DATA_ERR } sub=1)
  co0=0x1ce3910 ctrl=tcpv4 xprt=RAW mux=PASS data=STRM target=LISTENER:0xcb6c50
      flags=0x80203300 { CO_FL_XPRT_TRACKED | CO_FL_CONNECTED | CO_FL_ADDR_TO_SET | CO_FL_ADDR_FROM_SET | CO_FL_XPRT_READY | CO_FL_CTRL_READY } fd=1222 fd.state=22 { FD_EV_READY_R | FD_EV_READY_W } fd.cache=0 updt=0 fd.tmask=0x1
      cs=0x2176bf0 csf=0x00000200 { CS_FL_RCV_MORE } ctx=0x1d84150
  co1=0x21daa50 ctrl=tcpv4 xprt=SSL mux=PASS data=STRM target=SERVER:0xcb9420
      flags=0x00283306 { CO_FL_CONNECTED | CO_FL_SOCK_WR_SH | CO_FL_ADDR_TO_SET | CO_FL_ADDR_FROM_SET | CO_FL_XPRT_READY | CO_FL_CTRL_READY | CO_FL_CURR_RD_ENA | CO_FL_XPRT_RD_ENA } fd=1506 fd.state=25 { FD_EV_READY_W | FD_EV_POLLED_R | FD_EV_ACTIVE_R } fd.cache=0 updt=0 fd.tmask=0x1
      cs=0x1a68040 csf=0x00000240 { CS_FL_RCV_MORE | CS_FL_SHWN } ctx=(nil)
  req=0x2044e30 (f=0x48c4a820 { CF_EOI | CF_NEVER_WAIT | CF_AUTO_CONNECT | CF_READ_DONTWAIT | CF_WROTE_DATA | CF_AUTO_CLOSE | CF_SHUTW | CF_WRITE_ERROR | CF_SHUTR } an=0x0 pipe=0 tofwd=0 total=650)
      an_exp=<NEVER> rex=<NEVER> wex=<NEVER>
      buf=0x2044e38 data=(nil) o=0 p=0 req.next=0 i=0 size=0
  res=0x2044e90 (f=0x80408008 { CF_ISRESP | CF_READ_DONTWAIT | CF_AUTO_CLOSE | CF_READ_ERROR } an=0x0 pipe=0 tofwd=-1 total=0)
      an_exp=<NEVER> rex=? wex=<NEVER>
      buf=0x2044e98 data=(nil) o=0 p=0 rsp.next=0 i=0 size=0

What apparently happened here is the following sequence:

  - a client connected and sent a request
  - the request was forwarded to the server, causing a connection to be
    established
  - this connection doesn't immediately establish
  - the connection then establishes and the request is sent
  - the stream is notified and at the exact same instant the connect timeout expires
  - sess_update_st_con_tcp() notices there's SI_FL_EXP + CF_WROTE_DATA
    and sets the state to SI_ST_EST so that sess_establish() sets the
    analysers appropriately but sets an error on the channels
  - sess_establish sets up the analysers
  - analysers are called and leave since CF_*_ERROR are set
    (e.g. http_request_forward_body sees CF_WRITE_ERROR and switches to
     HTTP_MSG_ERROR then waits for other side to synchronize)
  - the timeouts are refreshed since CF_*_ERROR are part of CF_*_ACTIVITY.

From there, there will not be any I/O acitvity anymore, and thus no
ability to set CF_SHUTR on the response channel to turn the SI_ST_EST
to SI_ST_DIS then SI_ST_CLO. The timeouts not being refreshed cause
the stream to leave in expired state and to be instantly woken up
again, making no progress. And the timeouts cannot end this loop as
an apparent activity remains on the stream.

Note that the issue could not be reproduced even when cheating with
flags, because it involves too many bad luck situations. However,
unrolling the code by hand with the flags above definitely proves the
situation is stuck.

The real issue is to have a CF_READ_ERROR without CF_SHUTR on the
response channel, so we're missing a shutdown here. A simple shutw()
will close both sides since we're stlil in SI_ST_CON. This issue was
introduced in 1.5-dev13 as part of a fix for a closely related bug :
e3224e870 ("BUG/MINOR: session: ensure that we don't retry connection
if some data were sent"), but it is very likely that the issue could
not happen in earlier versions due to the different I/O sequences.
The bug disappeared from 2.0-dev6 with commit d66ed88a7 ("MEDIUM:
stream: re-arrange the connection setup status reporting") when the
SI_ST_RDY state was introduced because the shutdown call could be
placed there.

Since this issue only affects 1.9 it doesn't have an equivalent upstream
commit ID. It doesn't seem possible to trigger it before 1.9 so it would
be better not to backport it for no reason, unless a similar trace is
captured.